Split Service Provider to make VR Service Provider
