+++ Assets/LeapMotion/Scripts/LeapVRServiceProvider.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4bd4101d67c8f544986fc8551d1a0d4c
+timeCreated: 1494528272
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
